Move Livy and Celery functions out of views.py
