+++ core/views/__init__.py
@@ -12,3 +12,4 @@ from .record_identifer_transformation_scenario import *
 from .stateio import *
 from .core_background_tasks import *
 from .export import *
+from .external_background_tasks import *
